Add files via upload
Slide 6: Here we can see that the top line representing office supplies has the highest profit margin over the historical data, with an average of 68.37% profit margin. Technology comes in second, the middle line, coming in with an average of 25.86% profit margin. Lastly, we see that the Furniture category has the lowest profit margin over the historical data, with an average of 5.67%.
Slide 7: In this slide we see two bar charts: the blue chart on the left shows total profit over the last four years, while the orange chart on the right shows total quantity sold in that same time frame. We can see that both charts are ordered by total profit; technology has the highest profit, while furniture has the lowest. Yet, if we look at the orange chart on the right, they are out of order! Here, office supplies have the most quantity ordered, while technology has the lowest.
Slide 8: In this last analysis slide, we wanted to show the same data from the previous slide, total profit vs total quantity sold over the last four years, but we decided to break it down even further and look at the differences in each subcategory. This gives us a more in-depth look at which areas we should focus our suggestions for growth. By looking at these charts, the left side chart showing total profit and the right side showing quantity sold, we can see that technology holds the top three spots for highest total profit: the outlier here is the subcategory of machines, only profiting $3,387 over the past 4 years. The biggest concern our team has here is the furniture category: the subcategories chairs and furnishings had decent, mid-level profits over the 4 years, yet at the bottom (coming in with negative profits) we see bookcases and tables - If we look over at the right chart, they both had a decent count of quantity sold in that same time frame. Compared to that, copiers have  the highest total profit, and we see that this product subcategory has the least amount of quantity sold.
RECOMMENDATIONS
Slide 10: So if you see the top 10 product sales chart we have chairs and phones with the highest gross sales but when we deep dive into it we found that it is not as profitable as other products like copiers. There might be a few reasons like either we are giving too much discount like on average on chairs we are giving 17% & on Phones approx. 15.5%  discount or our profit ratio is much less on these products. 
Slide 11:If we see we have a 12% profit ratio on phones and just 4.5% on chairs. In order to increase the profit in the next 4 quarter cycle we want to test our hypothesis by doing A/B tests on our assumptions like the impact of decreasing the total discount and/or increasing the profit margin on these products to see overall profit in the next 4 quarters. 
Slide 12:Based on the historical data available and taking into account other factors for the next four quarters we predict the continuous positive trend, meaning we will be seeing overall growth in the year 2019 too. The p-value we calculated is 0.0089 which means our confidence level to achieve our predicted target is 99.11%.